+++ scripts/chain/shutdown_nodes.sh
@@ -27,22 +27,34 @@ if [ $FILES_EXISTS == "true" ]; then
     echo "---------- Stopping systemd service files --------"
     for (( a=1; a<=$NUM_VALS; a++ ))
     do
-        sudo -S systemctl stop $DAEMON-${a}.service
-        echo "-- Stopped $DAEMON-${a}.service --"
+        if [ -x "$(command -v systemctl)" ]; then
+            sudo -S systemctl stop $DAEMON-${a}.service
+            echo "-- Stopped $DAEMON-${a}.service -"
+            continue
+        fi
+
+        pid_path=$DAEMON_HOME-$a/pid
+        if [ -f "$pid_path" ]; then
+            pid_value=$(cat $pid_path)
+            kill -s 15 $pid_value
+            echo "-- Stopped $DAEMON-${a} by killing PID: $pid_value --"
+        fi
     done
     echo "------- Running unsafe reset all ---------"
     for (( a=1; a<=$NUM_VALS; a++ ))
     do
-        $DAEMON tendermint unsafe-reset-all  --home $DAEMON_HOME-$a    
+        $DAEMON tendermint unsafe-reset-all  --home $DAEMON_HOME-$a
         rm -rf $DAEMON_HOME-$a
-        echo "-- Executed $DAEMON unsafe-reset-all  --home $DAEMON_HOME-$a --"
+        echo "-- Executed $DAEMON unsafe-reset-all --home $DAEMON_HOME-$a --"
     done
-    echo "---------- Disabling systemd process files --------"
-    for (( a=1; a<=$NUM_VALS; a++ ))
-    do
-    sudo -S systemctl disable $DAEMON-${a}.service
-    echo "-- Executed sudo -S systemctl disable $DAEMON-${a}.service --"
-    done
+    if [ -x "$(command -v systemctl)" ]; then
+        echo "---------- Disabling systemd process files --------"
+        for (( a=1; a<=$NUM_VALS; a++ ))
+        do
+            sudo -S systemctl disable $DAEMON-${a}.service
+            echo "-- Executed sudo -S systemctl disable $DAEMON-${a}.service --"
+        done
+    fi
 else
     echo "----No simd services running-----"
 fi
